refactor: break key handlers into seperate files
